6topSublayer: ignore tx succ signals for 6top req
previously, also signals indicating tx success of 6top requests messages
were considered, so that a transaction would be marked as successfull
early leading to a schedule mismatch between the two nodes involved in
the transaction. as we are doing 2-step transactions, a requesting node
should wait for the 6top success response. a answering node waits for
the L2 acknowledgement of its 6top success response.